requirements.txt
kivy
cython
flask 
flask_sqlalchemy
pyyaml
python-dotenv
pyyaml
